Factor out MediaControls, and add loopback latency control
