+++ client/src/app/store.ts
@@ -1,10 +1,12 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
+import mediaBarReducer from '../features/mediaBar/mediaBarSlice';
 import musiciansReducer from '../features/musicians/musiciansSlice';
 import roomReducer from '../features/room/roomSlice';
 import recordingReducer from '../features/recording/recordingSlice';
 
 export const store = configureStore({
   reducer: {
+    mediaBar: mediaBarReducer,
     musicians: musiciansReducer,
     room: roomReducer,
     recording: recordingReducer,
