+++ client/src/features/audioInput/AudioStreamHooks.ts
@@ -3,6 +3,7 @@ import { useSelector } from 'react-redux';
 import { Base64 } from 'js-base64';
 
 import { usePatch } from '../feathers/FeathersHooks';
+import { selectLoopbackLatencyMs } from '../mediaBar/mediaBarSlice';
 import {
   selectMyTrackId,
   selectRecordingState,
@@ -184,19 +185,18 @@ export function useStreamData(stream: MediaStream | null): DataResponse {
   useChunkPoster(useSelector(selectMyTrackId), setWorkletCallback);
   const recordingState = useSelector(selectRecordingState);
   const delaySeconds = useSelector(selectRecordingDelaySeconds);
+  const loopbackLatencyMs = useSelector(selectLoopbackLatencyMs);
 
   // Hook to start the worklet when recordingState says so. Starting the audio
   // worklet is idempotent, so it's OK ifsend the message multiple times
   useEffect(
     () => {
       if (recordingState === 'recording') {
+        console.log(delaySeconds + loopbackLatencyMs / 1000);
         postWorkletMessage({
           action: 'start',
           // more negative -> delay mic more
-          // NOTE(gnewman): I found my laptop has a loopback delay of about 100
-          // ms, but we still need to implement calculating that easily in
-          // Cedar!
-          delaySeconds: delaySeconds + 0.1,
+          delaySeconds: delaySeconds + loopbackLatencyMs / 1000,
         });
       } else if (recordingState === 'stopped') {
         postWorkletMessage({ action: 'stop' });
